requirements.txt
asgiref==3.11.1
Django==6.0.3
djangorestframework-simplejwt==5.4.0
pillow==12.1.1
sqlparse==0.5.5
